+++ NEWS.md
@@ -1,3 +1,5 @@
+# admtools (development version)
+
 # admtools 0.4.0
 
 * added transformations for S3 class `pre_paleoTS` to integrate with `StratPal` and `paleoTS` package