+++ NEWS.md
@@ -1,4 +1,4 @@
-# admtools (development version)
+# admtools 0.5.0
 
 * new helper functions for estimation of age-depth models